Fixed incorrect melee damage stat shown in ui.

+++ src/server/game/Entities/Player/Player.cpp
@@ -2522,7 +2522,7 @@ void Player::InitStatsForLevel(bool reapplyMods)
     {
         SetUInt32Value(UF::ACTIVE_PLAYER_FIELD_MOD_DAMAGE_DONE_NEG + i, 0);
         SetUInt32Value(UF::ACTIVE_PLAYER_FIELD_MOD_DAMAGE_DONE_POS + i, 0);
-        SetUInt32Value(UF::ACTIVE_PLAYER_FIELD_MOD_DAMAGE_DONE_PCT + i, 1.0f);
+        SetFloatValue(UF::ACTIVE_PLAYER_FIELD_MOD_DAMAGE_DONE_PCT + i, 1.0f);
         SetUpdateFieldValue(m_values.ModifyValue(&Player::m_activePlayerData).ModifyValue(&UF::ActivePlayerData::ModDamageDoneNeg, i), 0);
         SetUpdateFieldValue(m_values.ModifyValue(&Player::m_activePlayerData).ModifyValue(&UF::ActivePlayerData::ModDamageDonePos, i), 0);
         SetUpdateFieldValue(m_values.ModifyValue(&Player::m_activePlayerData).ModifyValue(&UF::ActivePlayerData::ModDamageDonePercent, i), 1.0f);
